+++ hlx_statics/blocks/infocard/infocard.css
@@ -1,0 +1,1 @@
+@import url('../info-card/info-card.css');
