+++ hlx_statics/styles/styles.css
@@ -320,6 +320,7 @@ main div.github-actions-wrapper,
 main div.image-text-wrapper,
 main div.info-wrapper,
 main div.info-card-wrapper,
+main div.infocard-wrapper,
 main div.info-columns-wrapper,
 main div.inline-nested-alert-wrapper,
 main div.inlinealert-wrapper,
